Command palette and hosting config card share notice copy
